feat(build): Enable option to integrate nodesetLoader into open62541 v1.3.2 (#196)
