+++ .github/workflows/GccIntegrationTest.yml
@@ -5,6 +5,7 @@ on:
   push:
     branches:
       - master
+  pull_request:
   workflow_dispatch:
 
 env:
