+++ README.md
@@ -43,6 +43,7 @@ make
 
 ```c
 #include <NodesetLoader/backendOpen62541.h>
+#include <NodesetLoader/dataTypes.h>
 #include <open62541/plugin/log_stdout.h>
 #include <open62541/server.h>
 #include <open62541/server_config_default.h>
@@ -64,8 +65,10 @@ int main(int argc, const char *argv[]) {
   }
   UA_StatusCode retval = UA_Server_run(server, &running);
   //NodesetLoader is allocating memory for custom dataTypes, user has to manually clean up
-  cleanupCustomTypes(UA_Server_getConfig(server)->customDataTypes);
+  const UA_DataTypeArray *customTypes =
+    UA_Server_getConfig(server)->customDataTypes;
   UA_Server_delete(server);
+  NodesetLoader_cleanupCustomDataTypes(customTypes);
   return retval == UA_STATUSCODE_GOOD ? EXIT_SUCCESS : EXIT_FAILURE;
 }
 ```
