+++ backends/CMakeLists.txt
@@ -1,6 +1,42 @@
+# set backend variables in current scope
+set(NODESETLOADER_BACKEND_SOURCES "")
+set(NODESETLOADER_BACKEND_PUBLIC_INCLUDES "")
+set(NODESETLOADER_BACKEND_PRIVATE_INCLUDES "")
+set(NODESETLOADER_BACKEND_DEPS_LIBS "")
+set(NODESETLOADER_BACKEND_PUBLIC_HEADERS "")
+
 if(${ENABLE_BACKEND_STDOUT})
     add_subdirectory(stdout)
 endif()
+
 if(${ENABLE_BACKEND_OPEN62541})
     add_subdirectory(open62541)
+
+    list(APPEND NODESETLOADER_BACKEND_SOURCES
+        ${NODESETLOADER_BACKEND_OPEN62541_SOURCES})
+    list(APPEND NODESETLOADER_BACKEND_PUBLIC_INCLUDES
+        ${NODESETLOADER_BACKEND_OPEN62541_PUBLIC_INCLUDES})
+    list(APPEND NODESETLOADER_BACKEND_PRIVATE_INCLUDES
+        ${NODESETLOADER_BACKEND_OPEN62541_PRIVATE_INCLUDES})
+    list(APPEND NODESETLOADER_BACKEND_DEPS_LIBS
+        ${NODESETLOADER_BACKEND_OPEN62541_DEPS_LIBS})
+    list(APPEND NODESETLOADER_BACKEND_PUBLIC_HEADERS
+        ${NODESETLOADER_BACKEND_OPEN62541_PUBLIC_HEADERS})
 endif()
+
+# export backend variables in parent scope
+set(NODESETLOADER_BACKEND_SOURCES
+    ${NODESETLOADER_BACKEND_SOURCES}
+    PARENT_SCOPE)
+set(NODESETLOADER_BACKEND_PUBLIC_INCLUDES
+    ${NODESETLOADER_BACKEND_PUBLIC_INCLUDES}
+    PARENT_SCOPE)
+set(NODESETLOADER_BACKEND_PRIVATE_INCLUDES
+    ${NODESETLOADER_BACKEND_PRIVATE_INCLUDES}
+    PARENT_SCOPE)
+set(NODESETLOADER_BACKEND_DEPS_LIBS
+    ${NODESETLOADER_BACKEND_DEPS_LIBS}
+    PARENT_SCOPE)
+set(NODESETLOADER_BACKEND_PUBLIC_HEADERS
+    ${NODESETLOADER_BACKEND_PUBLIC_HEADERS}
+    PARENT_SCOPE)
