+++ backends/open62541/src/DataTypeImporter.c
@@ -229,7 +229,7 @@ static void setPaddingMemsize(UA_DataType *type,
             {
                 tm->padding = (UA_Byte)sizeof(UA_UInt32);
             }
-            else
+            else if (memberType->typeKind != UA_DATATYPEKIND_UNION)
             {
                 int align = getAlignment(memberType, customTypes);
 #ifdef USE_MEMBERTYPE_INDEX
@@ -559,9 +559,6 @@ static void calcMemSize(DataTypeImporter *importer)
         {
             // we can calculate the memsize if the memsize of all membertypes is
             // known
-#ifdef ENABLE_BACKEND_OPEN62541
-            setPaddingMemsize(type, importer->types);
-#else
             if (readyForMemsizeCalc(type, importer->types->types))
             {
                 setPaddingMemsize(type, importer->types);
@@ -570,7 +567,6 @@ static void calcMemSize(DataTypeImporter *importer)
             {
                 allTypesFinished = false;
             }
-#endif
         }
     }
 }
@@ -668,6 +664,9 @@ DataTypeImporter *DataTypeImporter_new(struct UA_Server *server)
         {
             return importer;
         }
+#ifndef USE_CLEANUP_CUSTOM_DATATYPES
+        newCustomTypes->cleanup = UA_TRUE;
+#endif
         newCustomTypes->next = config->customDataTypes;
         config->customDataTypes = newCustomTypes;
     }
