+++ backends/open62541/src/customDataType.c
@@ -31,6 +31,34 @@ const struct UA_DataType *findCustomDataType(const UA_NodeId *typeId,
     return NULL;
 }
 
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
+static void cleanupCustomTypes(const UA_DataTypeArray *types)
+{
+    while (types)
+    {
+        const UA_DataTypeArray *next = types->next;
+        if (types->types)
+        {
+            for (const UA_DataType *type = types->types;
+                 type != types->types + types->typesSize; type++)
+            {
+                free((void*)(uintptr_t)type->typeName);
+                for (UA_DataTypeMember *m = type->members;
+                     m != type->members + type->membersSize; m++)
+                {
+                    free((void*)(uintptr_t)m->memberName);
+                    m->memberName = NULL;
+                }
+                free((void*)type->members);
+            }
+        }
+        free((void*)(uintptr_t)types->types);
+        free((void*)(uintptr_t)types);
+        types = next;
+    }
+}
+#endif
+
 const struct UA_DataType *
 NodesetLoader_getCustomDataType(struct UA_Server *server,
                                 const UA_NodeId *typeId)
@@ -39,3 +67,11 @@ NodesetLoader_getCustomDataType(struct UA_Server *server,
     const UA_DataTypeArray *types = config->customDataTypes;
     return findCustomDataType(typeId, types);
 }
+
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
+void
+NodesetLoader_cleanupCustomDataTypes(const UA_DataTypeArray *customTypes)
+{
+    cleanupCustomTypes(customTypes);
+}
+#endif
