+++ backends/open62541/tests/dataTypeImport/enum.c
@@ -28,8 +28,14 @@ static void teardown(void)
 {
 
     UA_Server_run_shutdown(server);
-    cleanupCustomTypes(UA_Server_getConfig(server)->customDataTypes);
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
+    const UA_DataTypeArray *customTypes =
+        UA_Server_getConfig(server)->customDataTypes;
+#endif
     UA_Server_delete(server);
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
+    NodesetLoader_cleanupCustomDataTypes(customTypes);
+#endif
 }
 
 START_TEST(OperatingModeEnum)
