+++ backends/open62541/tests/dataTypeImport/subDataType.c
@@ -27,10 +27,14 @@ static void setup(void)
 static void teardown(void)
 {
     UA_Server_run_shutdown(server);
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
     const UA_DataTypeArray *customTypes =
-        UA_Server_getConfig(server)->customDataTypes; 
+        UA_Server_getConfig(server)->customDataTypes;
+#endif 
     UA_Server_delete(server);
-    cleanupCustomTypes(customTypes);
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
+    NodesetLoader_cleanupCustomDataTypes(customTypes);
+#endif
 }
 
 START_TEST(SubTypeOfInt32)
