+++ backends/open62541/tests/import.c
@@ -11,6 +11,7 @@
 
 #include "testHelper.h"
 #include <NodesetLoader/backendOpen62541.h>
+#include <NodesetLoader/dataTypes.h>
 
 UA_Server *server;
 char* nodesetPath=NULL;
@@ -24,8 +25,14 @@ static void setup(void) {
 
 static void teardown(void) {
     UA_Server_run_shutdown(server);
-    cleanupCustomTypes(UA_Server_getConfig(server)->customDataTypes);
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
+    const UA_DataTypeArray *customTypes =
+        UA_Server_getConfig(server)->customDataTypes;
+#endif
     UA_Server_delete(server);
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
+    NodesetLoader_cleanupCustomDataTypes(customTypes);
+#endif
 }
 
 START_TEST(Server_ImportNodeset) {
