+++ backends/open62541/tests/subDataTypes.c
@@ -11,6 +11,7 @@
 
 #include "testHelper.h"
 #include <NodesetLoader/backendOpen62541.h>
+#include <NodesetLoader/dataTypes.h>
 
 UA_Server *server;
 char *nodesetPath = NULL;
@@ -26,10 +27,14 @@ static void setup(void)
 static void teardown(void)
 {
     UA_Server_run_shutdown(server);
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
     const UA_DataTypeArray *customTypes =
         UA_Server_getConfig(server)->customDataTypes;
+#endif
     UA_Server_delete(server);
-    cleanupCustomTypes(customTypes);
+#ifdef USE_CLEANUP_CUSTOM_DATATYPES
+    NodesetLoader_cleanupCustomDataTypes(customTypes);
+#endif
 }
 
 START_TEST(Server_loadNodeset)
